requirements.txt
pytorch >= 1.6.0
torchvision = 0.7.0
numpy = 1.16.4
pandas = 0.23.4
openpyxl = 2.5.12